Create separated file to import a project
