+++ webapp/core/ExportProject.js
@@ -1,0 +1,15 @@
+"use strict";
+
+/**
+ * Class responsible to export a project
+ * @class ExportProject
+ * 
+ * 
+ * @property {object} json - Object with data to export.
+ * @property {object} callback - Function to callback the result.
+ */
+var ExportProject = function(json, callback){
+
+}
+
+module.exports = ExportProject;
